FEATURE:  Add manual validation for Production Deploy
Broke the production deployment out into two workflows. First, upon a
push or PR to master the production-plan workflow is executed. The
person deploying the environment will review the output of this plan.
Then, the deploying user will manually kick off the deployment via the
production-apply workflow. The user will be prompted to select the
branch (defaulting to master) and type in APPLY into a text field.  The
first job in the workflow will validate that the user's input matches
APPLY.  If the input does not match, we fail the job and do not
continue.